Get combined notify plugin working
